+++ data-raw/DATASET.R
@@ -5,6 +5,9 @@
 xlsx_path <- system.file("example-data/02_simplified.xlsx", package = "evavelo")
 evavelo_example <- read_evavelo(xlsx_path)
 
+## id_quest_mismatch
+mismatch_path <- system.file("example-data/03_id_quest_mismatch.xlsx", package = "evavelo")
+quest_mismatch_example <- read_evavelo(mismatch_path)
 
 ## Minimal expected column names for each worksheet
 comptage_colnames <- c(
@@ -26,5 +29,6 @@ enquete_colnames <- c(
 calendrier_colnames <- c("id_site_enq", "date_enq")
 
 
-usethis::use_data(evavelo_example, comptage_colnames, enquete_colnames, calendrier_colnames,
+usethis::use_data(evavelo_example, quest_mismatch_example,
+                  comptage_colnames, enquete_colnames, calendrier_colnames,
                   overwrite = TRUE, internal = TRUE)